+++ sync-for-payables/src/sdk/models/operations/deletebill.ts
@@ -8,19 +8,28 @@ import { AxiosResponse } from "axios";
 
 export class DeleteBillRequest extends SpeakeasyBase {
     /**
-     * Unique identifier for a bill
+     * Unique identifier for a bill.
      */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=billId" })
     billId: string;
 
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
+    /**
+     * Unique identifier for a connection.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
 }
 
 export class DeleteBillResponse extends SpeakeasyBase {
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -36,9 +45,15 @@ export class DeleteBillResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     pushOperation?: shared.PushOperation;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }