+++ sync-for-payables/src/sdk/sdk.ts
@@ -40,6 +40,7 @@ export type SDKProps = {
      * The security details required to authenticate the SDK
      */
     security?: shared.Security | (() => Promise<shared.Security>);
+
     /**
      * Allows overriding the default axios client used by the SDK
      */
@@ -67,8 +68,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "1.2.0";
-    genVersion = "2.125.1";
+    sdkVersion = "1.3.0";
+    genVersion = "2.159.2";
+    userAgent = "speakeasy-sdk/typescript 1.3.0 2.159.2 3.0.0 @codat/sync-for-payables";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);