+++ sync-for-payables/RELEASES.md
@@ -49,3 +49,13 @@ Based on:
 - [typescript v1.2.0] sync-for-payables
 ### Releases
 - [NPM v1.2.0] https://www.npmjs.com/package/@codat/sync-for-payables/v/1.2.0 - sync-for-payables
+
+## 2023-10-17 14:46:10
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payables.yaml
+- Speakeasy CLI 1.100.2 (2.159.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.3.0] sync-for-payables
+### Releases
+- [NPM v1.3.0] https://www.npmjs.com/package/@codat/sync-for-payables/v/1.3.0 - sync-for-payables